RND-347 add client.resources.get_file
A shorthand for getting a single file. Pretty much exposes a
`requests.get`, but with the credentials already in place.

Also use this method in other calls.

Also move StreamedResponse to utils so I can import it easily
for typehints.
